requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
dataclasses-json==0.6.1
ffmpeg==1.4
filelock==3.13.0
Flask==3.0.0
flask-sock==0.7.0
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.325
langsmith==0.0.53
llvmlite==0.41.1
load-dotenv==0.1.0
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
openai==0.28.1
openai-whisper==20230918
packaging==23.2
playwright==1.39.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyee==11.0.1
PyJWT==2.8.0
pyngrok==7.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
srt==3.5.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.3.3
torch==2.1.0
tqdm==4.66.1
twilio==8.10.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
vosk==0.3.44
websockets==12.0
Werkzeug==3.0.1
whisper==1.1.10
wsproto==1.2.0
yarl==1.9.2
